+++ nuclia/sdk/__init__.py
@@ -4,6 +4,7 @@ from .kb import NucliaKB  # noqa
 from .kbs import NucliaKBS  # noqa
 from .nua import NucliaNUA  # noqa
 from .nuas import NucliaNUAS  # noqa
+from .resource import NucliaResource  # noqa
 from .search import NucliaSearch  # noqa
 from .upload import NucliaUpload  # noqa
 from .zones import NucliaZones  # noqa