+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-## 1.1.5 (unreleased)
+## 1.1.6 (unreleased)
 
 
 - Nothing changed yet.
 
+
+## 1.1.5 (2023-07-26)
+
+- Support summary as resource attribute
+- Allow to create kb
+- Allow to set default account
 
 ## 1.1.4 (2023-07-21)